+++ apiaudio/api_resources/file.py
@@ -1,35 +1,10 @@
-from posixpath import basename
-import aflr
-from aflr.api_request import APIRequest
-import os
+from aflr.helper_classes import UploadableResource, ListableResource
 
 
-class File(APIRequest):
+class File(UploadableResource, ListableResource):
     custom_audio_resource_path = "/file/customaudio/"
     OBJECT_NAME = "file"
     resource_path = "/file"
-
-    @classmethod
-    def upload_audiofile(cls, file_path):
-        # get presigned URL
-        payload = open(file_path, "rb")
-        filename = os.path.basename(file_path)
-        headers = {"Content-Type": "audio/mpeg"}
-        url = cls._get_request(
-            path_param=cls.custom_audio_resource_path + "uploadurl?filename=" + filename
-        )
-
-        mediaId = url["mediaId"]
-        fileUploadUrl = url["fileUploadUrl"]
-
-        response = cls._put_request_fileupload(
-            url=fileUploadUrl, headers=headers, data=payload
-        )
-
-        response = {
-            "message": f"Success. Please make sure to save this mediaId : {mediaId}"
-        }
-        return response
 
     @classmethod
     def get_download_url(cls, metaId):
@@ -38,7 +13,7 @@ class File(APIRequest):
         )
 
     @classmethod
-    def search_audiofiles(cls, tags=None):
+    def list(cls, tags=None):
         if tags:
             return cls._get_request(
                 path_param=cls.custom_audio_resource_path + "tags?tags=" + tags
