+++ apiaudio/helper_classes.py
@@ -1,3 +1,6 @@
+import os
+from posixpath import basename
+
 from .api_request import APIRequest
 
 
@@ -66,3 +69,27 @@ class DownloadableResource(APIRequest):
             local_filenames.append(local_filename)
 
         return local_filenames
+
+
+class UploadableResource(APIRequest):
+    @classmethod
+    def upload_audiofile(cls, file_path):
+        # get presigned URL
+        payload = open(file_path, "rb")
+        filename = os.path.basename(file_path)
+        headers = {"Content-Type": "audio/mpeg"}
+        url = cls._get_request(
+            path_param=cls.custom_audio_resource_path + "uploadurl?filename=" + filename
+        )
+
+        mediaId = url["mediaId"]
+        fileUploadUrl = url["fileUploadUrl"]
+
+        response = cls._put_request_fileupload(
+            url=fileUploadUrl, headers=headers, data=payload
+        )
+
+        response = {
+            "message": f"Success. Please make sure to save this mediaId : {mediaId}"
+        }
+        return response
